+++ src/index.ts
@@ -19,6 +19,7 @@ import { dnsSteps } from './steps/resource-manager/dns';
 import { privateDnsSteps } from './steps/resource-manager/private-dns';
 import { containerRegistrySteps } from './steps/resource-manager/container-registry';
 import { serviceBusSteps } from './steps/resource-manager/service-bus';
+import { cdnSteps } from './steps/resource-manager/cdn';
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields: {
@@ -68,5 +69,6 @@ export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = 
     ...privateDnsSteps,
     ...containerRegistrySteps,
     ...serviceBusSteps,
+    ...cdnSteps,
   ],
 };